django-demo/pyproject.toml
[tool.poetry]
name = "django-demo"
version = "0.1.0"
description = ""
authors = ["Yogesh Ambhore <yogeshambhore45@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
